Reorganize modules in @glint/transform
