+++ packages/transform/__tests__/offset-mapping.test.ts
@@ -1,6 +1,6 @@
 import { rewriteModule, TransformedModule, rewriteDiagnostic } from '../src';
 import { stripIndent } from 'common-tags';
-import { Range, SourceFile } from '../src/transformed-module';
+import { Range, SourceFile } from '../src/template/transformed-module';
 import ts from 'typescript';
 import { assert } from '../src/util';
 import { GlintEnvironment } from '@glint/config/lib/environment';
